reporter_3/pyproject.toml
[project]
name = "reporter-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.4",
    "django-crispy-forms>=2.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "plotly>=6.2.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "requests>=2.32.4",
    "weasyprint>=66.0",
]
